+++ src/main/java/fuzs/betteranimationscollection/client/element/ModelElement.java
@@ -2,9 +2,13 @@ package fuzs.betteranimationscollection.client.element;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.google.common.collect.Sets;
 import fuzs.betteranimationscollection.BetterAnimationsCollection;
 import fuzs.betteranimationscollection.client.renderer.entity.layers.ILayerModelAccessor;
-import fuzs.betteranimationscollection.mixin.client.accessor.ILivingRendererAccessor;
+import fuzs.betteranimationscollection.mixin.client.accessor.LivingRendererAccessor;
+import fuzs.puzzleslib.config.ConfigManager;
+import fuzs.puzzleslib.config.option.OptionsBuilder;
+import fuzs.puzzleslib.config.serialization.EntryCollectionBuilder;
 import fuzs.puzzleslib.element.AbstractElement;
 import fuzs.puzzleslib.element.side.IClientElement;
 import net.minecraft.client.Minecraft;
@@ -12,12 +16,17 @@ import net.minecraft.client.renderer.entity.LivingRenderer;
 import net.minecraft.client.renderer.entity.layers.LayerRenderer;
 import net.minecraft.client.renderer.entity.model.EntityModel;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 @SuppressWarnings("unchecked")
 public abstract class ModelElement extends AbstractElement implements IClientElement {
@@ -25,11 +34,13 @@ public abstract class ModelElement extends AbstractElement implements IClientEle
     private final Minecraft mc = Minecraft.getInstance();
     private final Map<LivingRenderer<?, ?>, EntityModel<?>> rendererToOrigModel = Maps.newHashMap();
     private final List<LayerTransformer<?>> layerTransformers = Lists.newArrayList();
+    protected final List<ResourceLocation> defaultEntityBlacklist = Lists.newArrayList();
+
+    private Set<EntityType<?>> blacklistedEntities = Sets.newHashSet();
 
     @Override
     public void loadClient() {
 
-        // find all renderers which normally use the super class of our model, so we can exchange them
         this.switchModels();
     }
 
@@ -37,6 +48,26 @@ public abstract class ModelElement extends AbstractElement implements IClientEle
     public void unloadClient() {
 
         this.restoreModels();
+    }
+
+    @Override
+    public void setupClientConfig(OptionsBuilder builder) {
+
+        builder.define("Mob Blacklist", this.defaultEntityBlacklist.stream()
+                .map(ResourceLocation::toString)
+                .collect(Collectors.toList())).comment("Mob variants these model changes shouldn't be applied to.", EntryCollectionBuilder.CONFIG_STRING).sync(v -> {
+
+            Set<EntityType<?>> newEntityBlacklist = ConfigManager.deserializeToSet(v, ForgeRegistries.ENTITIES);
+            if (!this.blacklistedEntities.equals(newEntityBlacklist)) {
+
+                this.blacklistedEntities = newEntityBlacklist;
+                if (this.isEnabled() && this.isLoaded()) {
+
+                    this.restoreModels();
+                    this.switchModels();
+                }
+            }
+        });
     }
 
     protected abstract EntityModel<? extends LivingEntity> getEntityModel();
@@ -48,17 +79,18 @@ public abstract class ModelElement extends AbstractElement implements IClientEle
 
     private <T extends LivingEntity, M extends EntityModel<T>> void switchModels() {
 
-        this.mc.getEntityRenderDispatcher().renderers.values().forEach(renderer -> {
+        this.mc.getEntityRenderDispatcher().renderers.forEach((entityType, renderer) -> {
 
-            if (renderer instanceof LivingRenderer) {
+            if (!this.blacklistedEntities.contains(entityType) && renderer instanceof LivingRenderer) {
 
                 LivingRenderer<T, M> livingRenderer = ((LivingRenderer<T, M>) renderer);
                 M model = (M) this.getEntityModel();
+                // find all renderers which normally use the super class of our model, so we can exchange them
                 if (livingRenderer.getModel().getClass().isInstance(model)) {
 
                     BetterAnimationsCollection.LOGGER.info("Replaced {} with {} for {}", livingRenderer.getModel().getClass().getSimpleName(), model.getClass().getSimpleName(), livingRenderer.getClass().getSimpleName());
                     this.rendererToOrigModel.putIfAbsent(livingRenderer, livingRenderer.getModel());
-                    ((ILivingRendererAccessor<T, M>) livingRenderer).setModel(model);
+                    ((LivingRendererAccessor<T, M>) livingRenderer).setModel(model);
                     this.transformLayers(livingRenderer, true);
                 }
             }
@@ -70,7 +102,7 @@ public abstract class ModelElement extends AbstractElement implements IClientEle
         this.rendererToOrigModel.forEach((key, value) -> {
 
             BetterAnimationsCollection.LOGGER.info("Restored {} for {}", value.getClass().getSimpleName(), key.getClass().getSimpleName());
-            ((ILivingRendererAccessor<T, S>) key).setModel((S) value);
+            ((LivingRendererAccessor<T, S>) key).setModel((S) value);
             this.transformLayers(key, false);
         });
     }
@@ -82,7 +114,7 @@ public abstract class ModelElement extends AbstractElement implements IClientEle
             return;
         }
 
-        for (LayerRenderer<?, ?> layer : ((ILivingRendererAccessor<?, ?>) livingRenderer).getLayers()) {
+        for (LayerRenderer<?, ?> layer : ((LivingRendererAccessor<?, ?>) livingRenderer).getLayers()) {
 
             if (layer instanceof ILayerModelAccessor) {