+++ src/main/java/fuzs/betteranimationscollection/mixin/client/accessor/GhastModelAccessor.java
@@ -1,16 +1,13 @@
 package fuzs.betteranimationscollection.mixin.client.accessor;
 
 import com.google.common.collect.ImmutableList;
-import net.minecraft.client.renderer.entity.model.SquidModel;
+import net.minecraft.client.renderer.entity.model.GhastModel;
 import net.minecraft.client.renderer.model.ModelRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin(SquidModel.class)
-public interface ISquidModelAccessor {
-
-    @Accessor
-    ModelRenderer getBody();
+@Mixin(GhastModel.class)
+public interface GhastModelAccessor {
 
     @Accessor
     void setParts(ImmutableList<ModelRenderer> parts);